+++ pages/src/anchor/anchor-link.ts
@@ -1,7 +1,6 @@
 import {ESLMixinElement} from '../../../src/modules/esl-mixin-element/ui/esl-mixin-element';
-import {listen} from '../../../src/modules/esl-utils/decorators/listen';
+import {listen, memoize} from '../../../src/modules/esl-utils/decorators';
 import {afterNextRender} from '../../../src/modules/esl-utils/async/raf';
-import {memoize} from '../../../src/modules/esl-utils/decorators/memoize';
 import {EventUtils} from '../../../src/modules/esl-utils/dom/events/utils';
 
 /**
